+++ src/Akeneo/Connectivity/Connection/back/tests/Unit/spec/Infrastructure/Webhook/Client/GuzzleWebhookClientSpec.php
@@ -15,6 +15,7 @@ use Akeneo\Connectivity\Connection\Domain\Webhook\Model\WebhookEvent;
 use Akeneo\Connectivity\Connection\Infrastructure\Webhook\Client\GuzzleWebhookClient;
 use Akeneo\Connectivity\Connection\Infrastructure\Webhook\Client\Signature;
 use Akeneo\Connectivity\Connection\Infrastructure\Webhook\RequestHeaders;
+use Akeneo\Platform\Bundle\PimVersionBundle\VersionProviderInterface;
 use Akeneo\Platform\Component\EventQueue\Author;
 use Akeneo\Platform\Component\EventQueue\Event;
 use Akeneo\Platform\Component\EventQueue\EventInterface;
@@ -40,7 +41,8 @@ class GuzzleWebhookClientSpec extends ObjectBehavior
     public function let(
         SendApiEventRequestLogger $sendApiEventRequestLogger,
         EventsApiRequestLoggerInterface $eventsApiRequestLogger,
-        EventDispatcherInterface $eventDispatcher
+        EventDispatcherInterface $eventDispatcher,
+        VersionProviderInterface $versionProvider,
     ): void {
         $eventDispatcher->dispatch(Argument::any())->willReturn(Argument::type('object'));
         $this->beConstructedWith(
@@ -49,7 +51,9 @@ class GuzzleWebhookClientSpec extends ObjectBehavior
             $sendApiEventRequestLogger,
             $eventsApiRequestLogger,
             $eventDispatcher,
-            ['timeout' => 0.5, 'concurrency' => 1]
+            ['timeout' => 0.5, 'concurrency' => 1],
+            $versionProvider,
+            \getenv('PFID'),
         );
     }
 
@@ -62,9 +66,12 @@ class GuzzleWebhookClientSpec extends ObjectBehavior
     public function it_sends_webhook_requests_in_bulk(
         SendApiEventRequestLogger $sendApiEventRequestLogger,
         EventsApiRequestLoggerInterface $eventsApiRequestLogger,
-        EventDispatcherInterface $eventDispatcher
+        EventDispatcherInterface $eventDispatcher,
+        VersionProviderInterface $versionProvider,
     ): void {
         $eventDispatcher->dispatch(Argument::any())->willReturn(Argument::type('object'));
+        $versionProvider->getVersion()->willReturn('v20210526040645');
+
         $mock = new MockHandler(
             [
                 new Response(200, ['Content-Length' => 0]),
@@ -83,7 +90,9 @@ class GuzzleWebhookClientSpec extends ObjectBehavior
             $sendApiEventRequestLogger,
             $eventsApiRequestLogger,
             $eventDispatcher,
-            ['timeout' => 0.5, 'concurrency' => 1]
+            ['timeout' => 0.5, 'concurrency' => 1],
+            $versionProvider,
+            \getenv('PFID'),
         );
 
         $author = Author::fromNameAndType('julia', Author::TYPE_UI);
@@ -135,6 +144,13 @@ class GuzzleWebhookClientSpec extends ObjectBehavior
         $timestamp = (int)$request->getHeader(RequestHeaders::HEADER_REQUEST_TIMESTAMP)[0];
         $signature = Signature::createSignature('a_secret', $timestamp, $body);
         Assert::assertEquals($signature, $request->getHeader(RequestHeaders::HEADER_REQUEST_SIGNATURE)[0]);
+
+        $userAgent = 'AkeneoPIM/v20210526040645';
+        if (false !== \getenv('PFID')) {
+            $userAgent .= ' '.\getenv('PFID');
+        }
+
+        Assert::assertSame($userAgent, $request->getHeader(RequestHeaders::HEADER_REQUEST_USERAGENT)[0]);
 
         $eventDispatcher
             ->dispatch(Argument::allOf(
@@ -198,8 +214,11 @@ class GuzzleWebhookClientSpec extends ObjectBehavior
     public function it_logs_a_failed_events_api_request(
         SendApiEventRequestLogger $sendApiEventRequestLogger,
         EventsApiRequestLoggerInterface $eventsApiRequestLogger,
-        EventDispatcherInterface $eventDispatcher
+        EventDispatcherInterface $eventDispatcher,
+        VersionProviderInterface $versionProvider,
     ): void {
+        $versionProvider->getVersion()->willReturn('v20210526040645');
+
         $mock = new MockHandler(
             [
                 new Response(500, ['Content-Length' => 0]),
@@ -217,7 +236,9 @@ class GuzzleWebhookClientSpec extends ObjectBehavior
             $sendApiEventRequestLogger,
             $eventsApiRequestLogger,
             $eventDispatcher,
-            ['timeout' => 0.5, 'concurrency' => 1]
+            ['timeout' => 0.5, 'concurrency' => 1],
+            $versionProvider,
+            \getenv('PFID'),
         );
 
         $author = Author::fromNameAndType('julia', Author::TYPE_UI);
@@ -255,8 +276,11 @@ class GuzzleWebhookClientSpec extends ObjectBehavior
     public function it_does_not_send_webhook_request_because_of_timeout(
         SendApiEventRequestLogger $sendApiEventRequestLogger,
         EventsApiRequestLoggerInterface $debugLogger,
-        EventDispatcherInterface $eventDispatcher
+        EventDispatcherInterface $eventDispatcher,
+        VersionProviderInterface $versionProvider,
     ): void {
+        $versionProvider->getVersion()->willReturn('v20210526040645');
+
         $container = [];
         $history = Middleware::history($container);
 
@@ -269,7 +293,9 @@ class GuzzleWebhookClientSpec extends ObjectBehavior
             $sendApiEventRequestLogger,
             $debugLogger,
             $eventDispatcher,
-            ['timeout' => 0.5, 'concurrency' => 1]
+            ['timeout' => 0.5, 'concurrency' => 1],
+            $versionProvider,
+            \getenv('PFID'),
         );
 
         $author = Author::fromNameAndType('julia', Author::TYPE_UI);