+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - PIM-10915: Fix attribute with numeric code throw 500 error on product history
 - PIM-10889: Update Category updated date after setting a labels and show category filtered by updated date on API REST 
 - PIM-10919: Fix Cleaning Products with removed attributes using identifiers instead of uuids
+- PIM-10911: Add user-agent when sending an event
 
 ## Improvements